EDITOR-#199. Control panel websocket message interface
