+++ controller-server/constants/index.ts
@@ -1,31 +1,19 @@
-const SYNC = "sync";
-const UPLOAD_LED = "uploadLed";
-const UPLOAD_CONTROL = "uploadControl";
-const START = "start";
-const LOAD = "load";
-const PLAY = "play";
-const PAUSE = "pause";
-const STOP = "stop";
-const LIGTHCURRENTSTATUS = "lightCurrentStatus";
-const KICK = "kick";
-const SHUTDOWN = "shutdown";
-const REBOOT = "reboot";
-const TERMINATE = "terminate";
+// Command Type
+enum CommandType {
+  SYNC = "sync",
+  UPLOAD_LED = "uploadLed",
+  UPLOAD_CONTROL = "uploadControl",
+  LOAD = "load",
+  PLAY = "play",
+  PAUSE = "pause",
+  STOP = "stop",
+  LIGTHCURRENTSTATUS = "lightCurrentStatus",
+  KICK = "kick",
+  SHUTDOWN = "shutDown",
+  REBOOT = "reboot",
+  BOARDINFO = "boardInfo",
+  INIT = "init",
+  TEST = "test"
+}
 
-const COMMANDS = {
-  SYNC,
-  UPLOAD_LED,
-  UPLOAD_CONTROL,
-  START,
-  LOAD,
-  PLAY,
-  PAUSE,
-  STOP,
-  LIGTHCURRENTSTATUS,
-  KICK,
-  SHUTDOWN,
-  REBOOT,
-  TERMINATE,
-};
-
-export default COMMANDS;
+export { CommandType };
